finish implementing refactor
