--- packages/client-core/src/common/components/TouchGamepad/TouchGamepadProps.tsx
@@ -1,4 +1,0 @@
-export type TouchGamepadProps = {
-  layout?: string
-  onBoardingStep?: number | null
-}
