+++ packages/client-core/src/common/components/TouchGamepad/index.module.scss
@@ -5,6 +5,7 @@ $MobileControllButtonsSize: 3em;
   bottom: 110px;
   width: 50%;
   left: 0;
+  pointer-events: "auto";
 
   div {
     margin: auto;
