+++ packages/client-core/src/common/components/TouchGamepad/index.tsx
@@ -1,44 +1,71 @@
-import { useHookstate, useState } from '@hookstate/core'
-import React, { FunctionComponent, useEffect } from 'react'
+import { useHookstate } from '@hookstate/core'
+import React from 'react'
 import { Joystick } from 'react-joystick-component'
 
-import { useEngineState } from '@xrengine/engine/src/ecs/classes/EngineState'
-import { getComponent } from '@xrengine/engine/src/ecs/functions/ComponentFunctions'
+import { isTouchAvailable } from '@xrengine/engine/src/common/functions/DetectFeatures'
+import { isHMD } from '@xrengine/engine/src/common/functions/isMobile'
 import { GamepadAxis, GamepadButtons } from '@xrengine/engine/src/input/enums/InputEnums'
-import { InteractableComponent } from '@xrengine/engine/src/interaction/components/InteractableComponent'
 import { InteractState } from '@xrengine/engine/src/interaction/systems/InteractiveSystem'
 import { getState } from '@xrengine/hyperflux'
 
 import TouchAppIcon from '@mui/icons-material/TouchApp'
 
+import { AppState } from '../../services/AppService'
 import styles from './index.module.scss'
-import { TouchGamepadProps } from './TouchGamepadProps'
 
-export const TouchGamepad: FunctionComponent<TouchGamepadProps> = () => {
-  const triggerButton = (button: GamepadButtons, pressed: boolean): void => {
-    const eventType = pressed ? 'touchgamepadbuttondown' : 'touchgamepadbuttonup'
-    const event = new CustomEvent(eventType, { detail: { button } })
-    document.dispatchEvent(event)
+const triggerButton = (button: GamepadButtons, pressed: boolean): void => {
+  const eventType = pressed ? 'touchgamepadbuttondown' : 'touchgamepadbuttonup'
+  const event = new CustomEvent(eventType, { detail: { button } })
+  document.dispatchEvent(event)
+}
+
+const normalizeValues = (val) => {
+  const a = 1
+  const b = -1
+  const maxVal = 50
+  const minVal = -50
+
+  const newValue = (b - a) * ((val - minVal) / (maxVal - minVal)) + a
+
+  return newValue
+}
+
+const handleMove = (e) => {
+  const event = new CustomEvent('touchstickmove', {
+    detail: {
+      stick: GamepadAxis.LThumbstick,
+      value: { x: normalizeValues(-e.x), y: normalizeValues(e.y), angleRad: 0 }
+    }
+  })
+  document.dispatchEvent(event)
+}
+
+const handleStop = () => {
+  const event = new CustomEvent('touchstickmove', {
+    detail: { stick: GamepadAxis.LThumbstick, value: { x: 0, y: 0, angleRad: 0 } }
+  })
+  document.dispatchEvent(event)
+}
+
+const buttonsConfig: Array<{ button: GamepadButtons; label: string }> = [
+  {
+    button: GamepadButtons.A,
+    label: 'A'
   }
+]
 
+export const TouchGamepad = () => {
   const interactState = useHookstate(getState(InteractState))
   const availableInteractable = interactState.available.value?.[0]
+  const appState = useHookstate(getState(AppState))
 
-  const buttonsConfig: Array<{ button: GamepadButtons; label: string }> = [
-    {
-      button: GamepadButtons.A,
-      label: 'A'
-    }
-  ]
+  if (!isTouchAvailable || isHMD || appState.showTouchPad.value) return <></>
 
   const buttons = buttonsConfig.map((value, index) => {
     return (
       <div
         key={index}
-        className={
-          styles.controllButton + ' ' + styles[`gamepadButton_${value.label}`] + ' ' + styles.availableButton
-          // (hovered ? styles.availableButton : styles.notAvailableButton)
-        }
+        className={styles.controllButton + ' ' + styles[`gamepadButton_${value.label}`] + ' ' + styles.availableButton}
         onPointerDown={(): void => triggerButton(value.button, true)}
         onPointerUp={(): void => triggerButton(value.button, false)}
       >
@@ -46,34 +73,6 @@ export const TouchGamepad: FunctionComponent<TouchGamepadProps> = () => {
       </div>
     )
   })
-
-  const normalizeValues = (val) => {
-    const a = 1
-    const b = -1
-    const maxVal = 50
-    const minVal = -50
-
-    const newValue = (b - a) * ((val - minVal) / (maxVal - minVal)) + a
-
-    return newValue
-  }
-
-  const handleMove = (e) => {
-    const event = new CustomEvent('touchstickmove', {
-      detail: {
-        stick: GamepadAxis.LThumbstick,
-        value: { x: normalizeValues(-e.x), y: normalizeValues(e.y), angleRad: 0 }
-      }
-    })
-    document.dispatchEvent(event)
-  }
-
-  const handleStop = () => {
-    const event = new CustomEvent('touchstickmove', {
-      detail: { stick: GamepadAxis.LThumbstick, value: { x: 0, y: 0, angleRad: 0 } }
-    })
-    document.dispatchEvent(event)
-  }
 
   return (
     <>
@@ -92,5 +91,3 @@ export const TouchGamepad: FunctionComponent<TouchGamepadProps> = () => {
     </>
   )
 }
-
-export default TouchGamepad
