+++ packages/client-core/src/components/InstanceChat/index.module.scss
@@ -400,3 +400,19 @@
     top: 0;
   }
 }
+
+.modalConnecting {
+  right: 0;
+  bottom: 0;
+
+  // pointer-events: none;
+  z-index: 1000;
+  position: absolute;
+
+  .modalConnectingTitle {
+    padding: 20px 12px;
+    color: white;
+    text-shadow: 1px 1px 2px grey, 0 0 1em grey, 0 0 1em grey;
+    font-size: 0.7rem;
+  }
+}
