--- packages/client-core/src/components/Layout/index.module.scss
@@ -1,269 +1,0 @@
-body {
-  overflow: hidden;
-}
-
-.container {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-}
-
-.conferenceModeContainer {
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  z-index: 100000000;
-  flex-direction: column;
-  background-color: var(--mainBackground);
-
-  .toolbar {
-    display: flex;
-    padding: 3px;
-    flex-direction: row;
-    align-items: center;
-    justify-content: flex-end;
-    background-color: var(--navbarBackground);
-
-    .toolbarCrossButton {
-      display: flex;
-      flex-direction: row;
-      align-items: center;
-      margin: 1px;
-      padding: 1px;
-      cursor: pointer;
-      border-radius: 50%;
-      background-color: red;
-
-      svg {
-        width: 20px;
-        height: 20px;
-        color: white;
-
-        path {
-          fill: white;
-        }
-      }
-    }
-  }
-}
-
-.box {
-  display: flex;
-  width: 200px;
-  height: 150px;
-  margin: 10px;
-  border: solid black 2px;
-  text-align: center;
-
-  &:hover {
-    background-color: gray;
-    cursor: pointer;
-  }
-}
-
-.fullScreen {
-  position: fixed;
-  bottom: 20px;
-  left: 66.4px;
-
-  @media (max-width: 450px) {
-    left: 57.4px;
-  }
-}
-
-.expandMenu {
-  width: 50px;
-  height: 50px;
-  border: none;
-  color: white;
-  background: rgb(0 0 0 / 20%);
-  padding: 0;
-  position: absolute;
-  top: 10px;
-  left: 30px;
-  border-radius: 50%;
-  z-index: 11;
-
-  svg {
-    width: 50%;
-    height: 50%;
-  }
-
-  &.expanded {
-    background: var(--purpleButtonColor);
-  }
-
-  &::before {
-    content: " ";
-    position: absolute;
-    width: 50px;
-    height: 50px;
-    display: block;
-    transition: all 0.15s cubic-bezier(0.18, 0.89, 0.32, 1.28);
-    background: var(--purpleButtonColor);
-    top: 0;
-    left: 0;
-    z-index: -1;
-    border-radius: 50%;
-    transform: scale(0);
-  }
-
-  &:hover {
-    &::before {
-      transform: scale(1);
-    }
-  }
-
-  @media (min-width: 769px) {
-    display: none;
-  }
-
-  @media (max-width: 450px) {
-    width: 40px;
-    height: 40px;
-    left: 10px;
-
-    &::before {
-      width: 40px;
-      height: 40px;
-    }
-  }
-}
-
-.invite-toggle {
-  right: 0 !important;
-  top: 30px !important;
-  position: absolute !important;
-  z-index: 20;
-
-  button[aria-label="PersonAdd"] {
-    background: linear-gradient(30deg, #028000, #62ff00);
-    border: 2px solid white;
-  }
-}
-
-.showIconMedia {
-  position: fixed;
-  top: 10px;
-  left: 8px;
-}
-
-.showIcon {
-  position: fixed;
-  bottom: 20px;
-  left: 8px;
-}
-
-.rotate {
-  transform: rotateZ(-360deg);
-  transition: all 0.6s ease;
-}
-
-.rotateBack {
-  transform: rotateZ(0deg);
-  transition: all 0.6s ease;
-}
-
-.smBtn {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: black;
-  background: rgb(255 255 255 / 50%);
-  width: 40px;
-  height: 40px;
-  cursor: pointer;
-  z-index: 6;
-  border: none;
-  padding: 0;
-  border-radius: 50%;
-  opacity: 1;
-
-  &:hover {
-    box-shadow: 0 1rem 2rem rgba(#000, 0.2);
-    opacity: 1;
-  }
-
-  svg {
-    color: #313131;
-  }
-
-  @media (max-width: 450px) {
-    width: 30px;
-    height: 30px;
-  }
-}
-
-.btn {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: black;
-  background: white;
-  width: 50px;
-  height: 50px;
-  cursor: pointer;
-  z-index: 6;
-  border: none;
-  padding: 0;
-  border-radius: 50%;
-  opacity: 0.99;
-
-  &:hover {
-    box-shadow: 0 1rem 2rem rgba(#000, 0.2);
-    opacity: 1;
-  }
-
-  svg {
-    color: black;
-  }
-
-  @media (max-width: 450px) {
-    width: 40px;
-    height: 40px;
-  }
-}
-
-.refreshBtn {
-  position: fixed !important;
-  bottom: 20px !important;
-  left: 66.4px !important;
-
-  @media (max-width: 450px) {
-    left: 57.4px !important;
-  }
-}
-
-.respawn {
-  position: fixed;
-  bottom: 20px;
-  left: 125px;
-
-  @media (max-width: 450px) {
-    left: 106px;
-  }
-}
-
-@keyframes moveInBottom {
-  0% {
-    opacity: 0;
-    transform: translateY(1rem);
-  }
-
-  100% {
-    opacity: 1;
-    transform: translate(0);
-  }
-}
-
-@keyframes moveInTop {
-  0% {
-    opacity: 0;
-    transform: translateY(-1rem);
-  }
-
-  100% {
-    opacity: 1;
-    transform: translate(0);
-  }
-}
