+++ packages/client-core/src/components/LocationIcons/index.module.scss
@@ -1,0 +1,12 @@
+.rightSidebar {
+  display: flex;
+  flex-direction: column;
+  align-items: end;
+  justify-content: space-between;
+  position: fixed;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  pointer-events: none;
+  z-index: 10;
+}
