+++ packages/client-core/src/components/MediaIconsBox/index.module.scss
@@ -5,6 +5,7 @@
   z-index: 5;
   width: 100%;
   justify-content: center;
+  pointer-events: "auto";
 
   .iconContainer {
     background: rgb(0 0 0 / 20%);
