+++ packages/client-core/src/components/Shelves/useShelfStyles.ts
@@ -3,14 +3,12 @@ import { useHookstate } from '@hookstate/core'
 import { getState } from '@xrengine/hyperflux'
 
 import { AppState } from '../../common/services/AppService'
-import shelfStyles from './shelf.style.scss'
+import shelfStyles from './index.module.scss'
 
 export const useShelfStyles = () => {
   const appState = useHookstate(getState(AppState))
-  const showTopShelf = appState.showTopShelf.value
-  const showBottomShelf = appState.showBottomShelf.value
   return {
-    topShelfStyle: showTopShelf ? shelfStyles.animateTop : shelfStyles.fadeOutTop,
-    bottomShelfStyle: showBottomShelf ? shelfStyles.animateBottom : shelfStyles.fadeOutBottom
+    topShelfStyle: appState.showTopShelf.value ? shelfStyles.animateTop : shelfStyles.fadeOutTop,
+    bottomShelfStyle: appState.showBottomShelf.value ? shelfStyles.animateBottom : shelfStyles.fadeOutBottom
   }
 }
