--- packages/client-core/src/components/World/index.module.scss
@@ -1,30 +1,0 @@
-@import "@xrengine/client-core/src/styles/imports.module.scss";
-
-.modalConnecting {
-  right: 0;
-  bottom: 0;
-
-  // pointer-events: none;
-  z-index: 1000;
-  position: absolute;
-
-  .modalConnectingTitle {
-    padding: 20px 12px;
-    color: white;
-    text-shadow: 1px 1px 2px grey, 0 0 1em grey, 0 0 1em grey;
-    font-size: 0.7rem;
-  }
-}
-
-.rightSidebar {
-  display: flex;
-  flex-direction: column;
-  align-items: end;
-  justify-content: space-between;
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  pointer-events: none;
-  z-index: 10;
-}
