+++ packages/client/src/pages/offline/[locationName].tsx
@@ -2,13 +2,12 @@ import React, { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useRouteMatch } from 'react-router-dom'
 
-import Layout from '@xrengine/client-core/src/components/Layout'
 import { LoadingCircle } from '@xrengine/client-core/src/components/LoadingCircle'
+import { LocationIcons } from '@xrengine/client-core/src/components/LocationIcons'
 import { LoadEngineWithScene } from '@xrengine/client-core/src/components/World/LoadEngineWithScene'
 import OfflineLocation from '@xrengine/client-core/src/components/World/OfflineLocation'
 import { LocationAction } from '@xrengine/client-core/src/social/services/LocationService'
 import { DefaultLocationSystems } from '@xrengine/client-core/src/world/DefaultLocationSystems'
-import { Engine } from '@xrengine/engine/src/ecs/classes/Engine'
 import { useEngineState } from '@xrengine/engine/src/ecs/classes/EngineState'
 import { dispatchAction } from '@xrengine/hyperflux'
 
@@ -27,11 +26,12 @@ const LocationPage = () => {
   }, [])
 
   return (
-    <Layout>
+    <>
       {engineState.isEngineInitialized.value ? <></> : <LoadingCircle />}
       <LoadEngineWithScene injectedSystems={DefaultLocationSystems} />
       <OfflineLocation />
-    </Layout>
+      <LocationIcons />
+    </>
   )
 }
